requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
filelock
fsspec
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
joblib==1.5.2
markupsafe==3.0.3
mpmath==1.3.0
neo4j==6.0.3
networkx==3.6
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
urllib3==2.5.0